requirements.txt
langchain
flask
pydantic
pdfplumber
langchain_groq
langchain_huggingface
google-api-python-client
google-auth-oauthlib